requirements.txt
tensorflow==2.3.1
tensorflow-addons==0.12.1
absl-py==0.10.0
pandas==1.2.3
scikit-learn==0.23.2